Refactor the mssql connector to use base SqlConnector
