+++ lib/discovery.js
@@ -135,12 +135,12 @@ function mixinDiscovery(MsSQL) {
 
     var self = this;
     var calls = [function (callback) {
-      self.query(queryTables(options), callback);
+      self.execute(queryTables(options), callback);
     }];
 
     if (options.views) {
       calls.push(function (callback) {
-        self.query(queryViews(options), callback);
+        self.execute(queryViews(options), callback);
       });
     }
     async.parallel(calls, function (err, data) {
@@ -233,7 +233,7 @@ function mixinDiscovery(MsSQL) {
         cb(err, results);
       }
     };
-    this.query(sql, callback);
+    this.execute(sql, callback);
   };
 
   /*!
@@ -290,7 +290,7 @@ function mixinDiscovery(MsSQL) {
     cb = args.cb;
 
     var sql = queryForPrimaryKeys(owner, table);
-    this.query(sql, cb);
+    this.execute(sql, cb);
   };
 
   /*!
@@ -348,7 +348,7 @@ function mixinDiscovery(MsSQL) {
     cb = args.cb;
 
     var sql = queryForeignKeys(owner, table);
-    this.query(sql, cb);
+    this.execute(sql, cb);
   };
 
   /*!
@@ -393,7 +393,7 @@ function mixinDiscovery(MsSQL) {
     cb = args.cb;
 
     var sql = queryExportedForeignKeys(owner, table);
-    this.query(sql, cb);
+    this.execute(sql, cb);
   };
 
   function mysqlDataTypeToJSONType(mysqlType, dataLength) {
